requirements.txt
langchain>=0.1.14
langchain-openai>=0.1.1
openai>=1.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
langchain-community